STEP Bonus : 객체 미용 체조, 기능 추가
