+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Scenes/Invoice/InvoiceViewController.swift
@@ -6,8 +6,7 @@
 
 import UIKit
 
-class InvoiceViewController: UIViewController {
-
+final class InvoiceViewController: UIViewController {
     private let parcelInformation: ParcelInformation
     
     init(parcelInformation: ParcelInformation) {
