+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Scenes/Parcel/ParcelOrderViewController.swift
@@ -6,7 +6,7 @@
 
 import UIKit
 
-class ParcelOrderViewController: UIViewController, ParcelOrderViewDelegate {
+final class ParcelOrderViewController: UIViewController {
     private let parcelProcessor: ParcelProcessing
     
     init(parcelProcessor: ParcelProcessing) {
@@ -21,16 +21,17 @@ class ParcelOrderViewController: UIViewController, ParcelOrderViewDelegate {
         fatalError("init(coder:) has not been implemented")
     }
     
+    override func loadView() {
+        view = ParcelOrderView(delegate: self)
+    }
+}
+
+extension ParcelOrderViewController: ParcelOrderViewDelegate {
     func parcelOrderMade(_ parcelInformation: ParcelInformation) {
         parcelProcessor.process(parcelInformation: parcelInformation) { (parcelInformation) in
             let invoiceViewController: InvoiceViewController = .init(parcelInformation: parcelInformation)
             navigationController?.pushViewController(invoiceViewController, animated: true)
         }
     }
-    
-    override func loadView() {
-        view = ParcelOrderView(delegate: self)
-    }
-
 }
